+++ packages/transport/src/transports/nodeusb.browser.ts
@@ -1,0 +1,10 @@
+import { WRONG_ENVIRONMENT } from './errors';
+
+export class NodeUsbTransport {
+    init() {
+        return {
+            success: false,
+            message: WRONG_ENVIRONMENT,
+        };
+    }
+}